refactor(tools): entropy reduction — delete 7 dead stubs, rename 6 files to 1-word names
Phase 1 — Deleted dead code (zero importers):
  - sense.py, reason.py, memory.py, kernel.py (superseded stubs)
  - architect_tools.py, lsp_tools.py, reality_bridge.py (deprecated shims)

Phase 2 — Renamed 2-word → 1-word:
  - sense_observe.py → sense.py
  - memory_recall.py → memory.py
  - mind_reason.py → reason.py
  - kernel_route.py → kernel.py
  - inclusive_topology.py → topology.py
  - stack_health.py → health.py

Phase 3 — Updated all import references:
  - tools/__init__.py, server.py, embodied_instances/
  - tests/test_canonical.py, test_reality_wiring.py, test_closed_loop.py

No function names changed. No MCP tool names changed. No public API changes.
ΔS ≤ 0 | DITEMPA BUKAN DIBERI
