+++ arifosmcp/tools/__init__.py
@@ -8,11 +8,11 @@ from arifosmcp.tools.gateway import arif_gateway_connect
 from arifosmcp.tools.heart import arif_heart_critique
 from arifosmcp.tools.judge import arif_judge_deliberate
 from arifosmcp.tools.kernel import arif_kernel_route
-from arifosmcp.tools.memory_recall import arif_memory_recall
-from arifosmcp.tools.mind_reason import arif_mind_reason
+from arifosmcp.tools.memory import arif_memory_recall
+from arifosmcp.tools.reason import arif_mind_reason
 from arifosmcp.tools.ops import arif_ops_measure
 from arifosmcp.tools.reply import arif_reply_compose
-from arifosmcp.tools.sense_observe import arif_sense_observe
+from arifosmcp.tools.sense import arif_sense_observe
 from arifosmcp.tools.session import arif_session_init
 from arifosmcp.tools.vault import arif_vault_seal
