+++ arifosmcp/tools/health.py
