+++ arifosmcp/tools/topology.py
