+++ tests/test_closed_loop.py
@@ -18,7 +18,7 @@ import asyncio
 
 def test_reality_bridge():
     """Test Reality Bridge exists and works."""
-    from arifosmcp.tools.reality_bridge import RealityBridge
+    from arifosmcp.tools.reality import RealityBridge
 
     bridge = RealityBridge()
 
@@ -115,7 +115,7 @@ async def test_skill_execution():
 async def test_all_skills_wired():
     """Test all 9 skills accept reality_bridge parameter."""
     from skills import list_skills, SKILL_REGISTRY
-    from arifosmcp.tools.reality_bridge import RealityBridge
+    from arifosmcp.tools.reality import RealityBridge
 
     bridge = RealityBridge()
     skills = list_skills()
